requirements.txt
ccxt>=4.1.59
python-telegram-bot>=20.3
python-dotenv>=1.0.0
numpy==1.24.3
pandas==2.0.3
python-binance==1.0.19
python-telegram-bot==20.3
